move lively wallpaper js logic to the viewmodel
